requirements.txt
streamlit
pandas
numpy
numpy-financial
plotly
google-genai